Refactored code for player input.
I refactored and rearranged the code that deals with player input as to
make it easier to implement various commands a player can execute
in addition to just moving a chess piece. For that purpose, code that
dealt with input has been pulled together into game class and a
class variable has been included that stores all available commands.